requirements.txt
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
pyparsing==2.1.8
python-ceilometerclient==2.5.0
python-cinderclient==1.8.0
python-dateutil==2.5.3
python-designateclient==2.3.0
python-gitlab==0.17
python-glanceclient==2.5.0
python-heatclient==1.4.0
python-ironicclient==1.7.0
python-keystoneclient==3.4.0
python-magnumclient==2.3.0
python-mistralclient==2.1.1
python-neutronclient==5.1.0
python-nmap==0.6.1
python-novaclient==5.1.0
python-openstackclient==3.0.1
python-swiftclient==3.0.0
python-troveclient==2.5.0
pytimeparse==1.1.5
pytz==2016.10
PyYAML==3.11
requests==2.12.4
requestsexceptions==1.1.3
retrying==1.3.3
rfc3986==0.4.1
setproctitle==1.1.10
shade==1.11.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
stevedore==1.17.1
trollius==2.1
unicodecsv==0.14.1
vboxapi==1.0
voluptuous==0.9.3
waitress==0.9.0
warlock==1.2.0
WebOb==1.6.1
WebTest==2.0.23
Werkzeug==0.11.10
wrapt==1.10.8
